+++ client/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 All notable changes to the "Harbour and xHarbour" extension will be documented in this file.
 
+# 0.9.7
+ - missing files
+ 
 # 0.9.6
  - **server** [better outline and breadcump](https://github.com/APerricone/harbourCodeExtension/raw/master/images/0_9_6.png)
  - **debugger** fixed compilation with xHarbour, see #38